+++ textrun-workspace/features/file-content.feature
@@ -1,42 +1,32 @@
 Feature: verifying file content
 
-  As a documentation writer
-  I want to be able to verify created files
-  So that I am sure my tools performs the correct activities on the hard drive.
-
-  - the "verifyWorkspaceFileContent" action verifies files in the current workspace
-  - the filename is provided as strong text or italic text
-  - the expected content is provided as a code or fenced block
-
-
   Background:
     Given the workspace contains the file "hello.txt" with content:
       """
       Hello world!
       """
 
-
   Scenario: specify file name via emphasized text and content via code block
     Given the source code contains the file "01.md" with content:
       """
-      <a textrun="verify-workspace-file-content">
+      <a textrun="workspace/file-content">
 
       _hello.txt_ with content `Hello world!`
+
       </a>
       """
     When running text-run
     Then it signals:
-      | FILENAME | 01.md                    |
-      | LINE     | 1                        |
-      | MESSAGE  | verifying file hello.txt |
-
+      | FILENAME | 01.md                            |
+      | LINE     | 1                                |
+      | MESSAGE  | verify content of file hello.txt |
 
   Scenario: specify file name via strong text and content via fenced block
     Given the source code contains the file "01.md" with content:
       """
-      <a textrun="verify-workspace-file-content">
+      <a textrun="workspace/file-content">
 
-      __hello.txt__
+      **hello.txt** with content:
 
       ```
       Hello world!
@@ -45,15 +35,15 @@ Feature: verifying file content
       """
     When running text-run
     Then it signals:
-      | FILENAME | 01.md                    |
-      | LINE     | 1                        |
-      | MESSAGE  | verifying file hello.txt |
+      | FILENAME | 01.md                            |
+      | LINE     | 1                                |
+      | MESSAGE  | verify content of file hello.txt |
 
 
   Scenario: file content mismatch
     Given the source code contains the file "01.md" with content:
       """
-      <a textrun="verify-workspace-file-content">
+      <a textrun="workspace/file-content">
 
       __hello.txt__
 
@@ -73,7 +63,7 @@ Feature: verifying file content
   Scenario: file is missing
     Given the source code contains the file "01.md" with content:
       """
-      <a textrun="verify-workspace-file-content">
+      <a textrun="workspace/file-content">
 
       __zonk.txt__
 
@@ -82,7 +72,7 @@ Feature: verifying file content
       """
     When trying to run text-run
     Then the test fails with:
-      | FILENAME      | 01.md                   |
-      | LINE          | 1                       |
-      | ERROR MESSAGE | file zonk.txt not found |
-      | EXIT CODE     | 1                       |
+      | FILENAME      | 01.md                    |
+      | LINE          | 1                        |
+      | ERROR MESSAGE | file not found: zonk.txt |
+      | EXIT CODE     | 1                        |
